requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
appnope=0.1.4=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
blas=1.0=openblas
bottleneck=1.4.2=py310hbda83bc_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.11.26=hca03da5_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
comm=0.2.2=pyhd8ed1ab_0
debugpy=1.6.7=py310h313beb8_0
decorator=5.1.1=pyhd8ed1ab_1
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pyha770c72_1
importlib-resources=6.4.5=pypi_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.30.0=pyh707e725_0
jedi=0.19.2=pyhd8ed1ab_1
joblib=1.4.2=py310hca03da5_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
libsodium=1.0.18=h27ca646_1
llvm-openmp=14.0.6=hc6e5704_0
lxml=5.3.0=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
nibabel=5.3.2=pypi_0
nilearn=0.11.0=pypi_0
numexpr=2.10.1=py310h5d9532f_0
numpy=1.26.4=py310h3b2db8e_0
numpy-base=1.26.4=py310ha9811e2_0
openssl=3.4.0=h39f12f2_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py310hcf29cfe_0
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pip=24.2=py310hca03da5_0
platformdirs=4.3.6=pyhd8ed1ab_1
prompt-toolkit=3.0.48=pyha770c72_1
psutil=5.9.0=py310h1a28f6b_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_1
python=3.10.15=hb885b13_1
python-dateutil=2.9.0.post0=pyhff2d567_1
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py310hca03da5_0
pyzmq=25.1.2=py310h313beb8_0
readline=8.2=h1a28f6b_0
requests=2.32.3=pypi_0
scikit-learn=1.5.1=py310h46d7db6_0
scipy=1.14.1=py310hd336fd7_0
setuptools=75.1.0=py310hca03da5_0
six=1.16.0=pyhd8ed1ab_1
sqlite=3.45.3=h80987f9_0
stack_data=0.6.2=pyhd8ed1ab_0
threadpoolctl=3.5.0=py310h33ce5c2_0
tk=8.6.14=h6ba3021_0
tornado=6.4.1=py310h80987f9_0
tqdm=4.66.5=py310h33ce5c2_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.44.0=py310hca03da5_0
xz=5.4.6=h80987f9_1
zeromq=4.3.5=h313beb8_0
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h18a0788_1
